+++ src/infra/usecases/db/postgres/repositories/tasks/pg-tasks-repository.ts
@@ -1,6 +1,6 @@
 import { IGetAllTasksRepository } from "@data/protocols/db/task/get-all-tasks-repository";
 import { ITask } from "@domain/protocols/task";
-import { PgHelper } from "../helpers/pg-helper";
+import { PgHelper } from "../../helpers/pg-helper";
 import { IGetTasksByCategoryRepository } from "@data/protocols/db/task/get-tasks-by-category-repository";
 import { TaskByCategory } from "@domain/protocols/task-by-category";
 import { GetTasksByCategoryModel } from "@domain/usecases/tasks/get-tasks-by-category";
