+++ src/main/factory/tasks/get/get-tasks-by-category.ts
@@ -1,5 +1,5 @@
 import { DbGetByTasks } from '@data/usecases/db/tasks/get/db-get-by-tasks'
-import { PgTasksRepository } from '@infra/usecases/db/postgres/tasks/pg-tasks-repository'
+import { PgTasksRepository } from '@infra/usecases/db/postgres/repositories/tasks/pg-tasks-repository'
 import { GetTasksByCategoryController } from '@presentation/controller/tasks/get-by/get-tasks-by-category-controller'
 import { makeGetTasksByCategoryValidation } from './get-tasks-by-category-validation'
